[ contacts ] Simplify and extend fill algorithm
New uses the fill algorithm to identify subcontacts within
a particular support region.

The fill algorithm has been generalized to take function inputs
for "accepting" and "claiming" points it finds.

This means it can be used to both identify support points for fitting
as well as distinct positive curvature peaks within a support.

This algorithm seems to work reasonably well, but it does not cover
cases when the peaks are very close, for this we may have to use an
alternative curvature partitioning algorithm, (which looks for negative
curvature saddles, and uses Euler's theorem to find the proper dividing
line between the two points).

Signed-off-by: Martin Lueker-Boden <martin.lueker@gmail.com>